+++ packages/core/src/Switch.js
@@ -19,17 +19,30 @@ export const BEM = {
     iconWrapper: ROOT_BEM.element('icon-wrapper'),
 };
 
+/**
+ * <Switch>
+ * ===
+ *
+ * A `<Switch>` is a row component which can be turned either on or off,
+ * with its underlying `<input type=checkbox>`.`
+ */
+
 class Switch extends PureComponent {
     static propTypes = {
         /**
-         * Use `input` to inject props to the underlying <input>
+         * Use `input` to inject props to the underlying `<input>`
          */
         input: PropTypes.object, // eslint-disable-line react/forbid-prop-types
         checked: PropTypes.bool,
         defaultChecked: PropTypes.bool,
 
-        // <input type="checkbox" /> props
+        /**
+         * `<input type="checkbox" />` props
+         */
         disabled: PropTypes.bool,
+        /**
+         * `<input type="checkbox" />` props
+         */
         onChange: PropTypes.func,
     };