+++ packages/storybook/.storybook/main.js
@@ -1,0 +1,11 @@
+module.exports = {
+    stories: [
+        '../examples/**/*.stories.(js|mdx)',
+        '../examples/**/index.js',
+    ],
+
+    addons: [
+        '@storybook/addon-actions',
+        '@storybook/addon-docs',
+    ],
+};