+++ packages/storybook/examples/core/Tag.stories.js
@@ -1,0 +1,32 @@
+import React from 'react';
+
+import Tag from '@ichef/gypcrete/src/Tag';
+
+export default {
+    title: '@ichef/gypcrete|Tag',
+    component: Tag,
+};
+
+export function BasicTagExample() {
+    return (
+        <Tag>tag</Tag>
+    );
+}
+
+export function WithParentColor() {
+    return (
+        <div>
+            <span style={{ color: 'blue' }}>
+                Blue text
+                {' '}
+                <Tag>Tag</Tag>
+            </span>
+
+            <span style={{ color: 'red' }}>
+                Red text
+                {' '}
+                <Tag>Tag</Tag>
+            </span>
+        </div>
+    );
+}